+++ src/index.js
@@ -1,16 +1,21 @@
-#!/usr/bin/env node
+// @flow
 
 import chalk from 'chalk';
-import {ConfigType} from './types/config';
+import {type ConfigType} from './types/config';
 import cosmiconfig from 'cosmiconfig';
-import joi from 'joi';
+import Deploy from './Deploy';
+import joi from '@hapi/joi';
+import pkg from '../package';
 import program from 'commander';
-import pkg from '../package.json';
-import Run from './run';
 
 // Define global CLI options
 program
 	.version(pkg.version)
+	.option(
+		'-D, --dir <path>',
+		'Path to the directory you want to run your deployment from. Defaults to process.cwd().',
+		process.cwd()
+	)
 	.option('-v, --verbose', 'Show verbose debug messages')
 	.parse(process.argv);
 
@@ -38,22 +43,25 @@ const explorer = cosmiconfig('deploy');
 
 // Search for a configuration by walking up directories.
 // See documentation for search, below.
-explorer.search()
+explorer.search(program.dir)
 	.then(({
 		config,
 		filepath,
-		isEmpty
+		isEmpty,
 	}: {
 		config: ConfigType,
 		filepath: string,
-		isEmpty: boolean
+		isEmpty: boolean,
 	}): any => {
+		// eslint-disable-next-line no-console
 		if (isEmpty) return console.error(chalk.red(
 			`Configuration file ${filepath} is empty. Can't proceed.`
 		));
 
 		// Validate configuration
 		const result = joi.validate(config, configSchema);
+
+		// eslint-disable-next-line no-console
 		if (result.error) return console.error(chalk.red(
 			result.error.message
 		));
@@ -61,22 +69,27 @@ explorer.search()
 		program
 			.command('run')
 			.description('Perform a production deployment')
-			.action(() => new Run(config, program));
+			.action((): any => new Deploy(config, program));
 
 		// Warn about commands we don't support
 		program
 			.command('*')
-			.action((cmd) => {
+			.action((cmd: string) => {
+				// eslint-disable-next-line no-console
 				console.error(chalk.red(`'${cmd}' is not a valid command.`));
 			});
 
 		// Warn about missing commands
 		if (!program.args.length) {
+			// eslint-disable-next-line no-console
 			console.error(chalk.red(`No commands specified. Try 'deploy run' instead.`));
 		}
 
 		return program.parse(process.argv);
 	})
-	.catch((error) => console.error(chalk.red(
-		`Unexpected error occured: ${error.message}`
-	)));
+	.catch((error: Error) => {
+		// eslint-disable-next-line no-console
+		console.error(chalk.red(
+			`Unexpected error occured: ${error.message}`
+		));
+	});