+++ src/types/config.js
@@ -1,7 +1,11 @@
+// @flow
+
 export type ConfigType = {
 	host: string,
 	private_key_path: string,
 	sources: Array<string>,
 	target_dir: string,
 	user: string,
-}
+};
+
+export type CmdResponseType = [?string, ?string];